chore: добавлено игнорирование папки .tests в .gitignore
refactor: обновлены методы обработки вариаций товаров в ProcessingProducts.php
* изменено обновление вариаций, убрано скрытие существующих
* добавлен метод deleteGoods для удаления товаров и их вариаций
* улучшена обработка ошибок и валидация в RestV1M2M.php
* добавлены новые методы для работы с вариациями в RestConfig.php

+++ packages/WeppsExtensions/Addons/Rest/Rest.php
@@ -363,7 +363,7 @@ class Rest
 
 			return $handler->{$config['method']}($this->data);
 		} catch (\Exception $e) {
-			$status = $e->getCode() ?: 400;
+			$status = (int) ($e->getCode() ?: 400);
 			$this->status = $status;
 			return ['status' => $status, 'message' => $e->getMessage(), 'data' => null];
 		}
@@ -423,7 +423,7 @@ class Rest
 		$bearer = $jwt->bearer();
 
 		if ($bearer['status'] != 200) {
-			throw new \Exception('Authentication failed: ' . $bearer['message'], $bearer['status']);
+			throw new \Exception('Authentication failed: ' . $bearer['message'], (int) $bearer['status']);
 		}
 
 		if ($bearer['payload']['typ'] != 'auth') {
